requirements.txt
anthropic>=0.40.0
opencv-python>=4.9.0
python-dotenv>=1.0.0
requests>=2.32.0
onvif-zeep>=0.2.12
